backend/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai
python-dotenv
python-multipart